simplify the ispyb code to not use the proxy but to use a pip packaged version of the api
